Neatened some debug messages. Changed main.cpp so ScEd can use it. Fixed minor inconsistencies and warnings.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1708 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/sced/ui/UnitToolsDlgBar.cpp
@@ -156,7 +156,7 @@ void CUnitToolsDlgBar::OnButtonEdit()
 	
 	// get object at position
 	const std::vector<CObjectEntry*>& objects=g_ObjMan.m_ObjectTypes[GetCurrentObjectType()].m_Objects;
-	CObjectEntry* obj=objects[(int) pos-1];
+	CObjectEntry* obj=objects[(intptr_t)pos-1];
 	CMainFrame* mainfrm=(CMainFrame*) AfxGetMainWnd();
 	mainfrm->OnObjectProperties(obj);
 }
@@ -169,7 +169,7 @@ void CUnitToolsDlgBar::OnClickListObjectBrowser(NMHDR* pNMHDR, LRESULT* pResult)
 	// deselect current selection in list ctrl, if any
 	POSITION pos=listctrl->GetFirstSelectedItemPosition();
 	if (pos) {
-		g_ObjMan.SetSelectedObject(g_ObjMan.m_ObjectTypes[GetCurrentObjectType()].m_Objects[(int) pos-1]);
+		g_ObjMan.SetSelectedObject(g_ObjMan.m_ObjectTypes[GetCurrentObjectType()].m_Objects[(intptr_t)pos-1]);
 	}
 
 	// shift to add mode
